new credential pattern, for #178
